extract util functions
